requirements/base.txt
# Core dependencies for Task 1
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Data validation (if needed)
pydantic>=2.0.0,<3.0.0

